requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.32
alembic==1.13.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pymupdf==1.24.10
pdfminer.six==20240706
pillow==10.4.0
numpy==1.26.4
paddleocr==2.7.3
opencv-python-headless==4.8.1.78
rapidfuzz==3.9.7
selenium==4.25.0
webdriver-manager==4.0.2
beautifulsoup4==4.12.3
requests==2.32.3
nltk==3.9.1
pytesseract==0.3.13
spacy==3.7.6
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
PyJWT==2.10.1
email-validator==2.2.0
pytest==8.3.3
